Refactor: 디렉토리init을 Repository로 옮김.
